I hope open source doesn't see this

+++ src/ros/cameras/cameras3/src/properties/capsfilters.cpp
@@ -20,8 +20,8 @@ void set_scalefilter(GstElement* filter, const std::string format, const int wid
   GstCaps *caps = gst_caps_new_simple(
       mime.c_str(),
       "format", G_TYPE_STRING, format.c_str(),
-      "width", G_TYPE_INT, width/downscale,
-      "height", G_TYPE_INT, height/downscale,
+      "width", G_TYPE_INT, (int) ((float) width/ (float) downscale),
+      "height", G_TYPE_INT, (int) ((float) height/ (float) downscale),
       "framerate", GST_TYPE_FRACTION, framerate, framerate_denominator*downrate,
       "brightness", G_TYPE_INT, brightness,
       "contrast", G_TYPE_INT,  contrast,
@@ -29,3 +29,4 @@ void set_scalefilter(GstElement* filter, const std::string format, const int wid
   g_object_set(filter, "caps", caps, NULL);
   gst_caps_unref(caps);
 }
+
